+++ package.json
@@ -10,6 +10,8 @@
 		"lint": "eslint ."
 	},
 	"dependencies": {
+		"@react-navigation/bottom-tabs": "^6.0.9",
+		"@react-navigation/native": "^6.0.6",
 		"@types/uuid": "^8.3.3",
 		"axios": "^0.24.0",
 		"buffer": "^6.0.3",
@@ -20,6 +22,8 @@
 		"react-native-hce": "^0.1.2",
 		"react-native-linear-gradient": "^2.5.6",
 		"react-native-nfc-manager": "^3.11.0",
+		"react-native-safe-area-context": "^3.3.2",
+		"react-native-screens": "^3.9.0",
 		"react-native-svg": "^12.1.1",
 		"react-native-version-info": "^1.1.1",
 		"react-native-webview": "^11.14.3",
