+++ src/components/fire.tsx
@@ -1,34 +1,24 @@
 import React from 'react';
-import {View, Text, StyleSheet, Image} from 'react-native';
-import LinearGradient from 'react-native-linear-gradient';
+import {View, Text, StyleSheet} from 'react-native';
 import Torch from '../assets/torch.svg';
 import TorchOff from '../assets/torchOff.svg';
 
-const FireView = (props: {fire: boolean}) => {
-	if (props.fire) {
-		return (
-			<View style={styles.container}>
-				<View style={styles.headlineBox}>
-					<Text style={styles.headlineText}>SHARE YOUR FIRE!</Text>
-				</View>
-				<View style={styles.fire}>
-					<Torch width={'100%'} height={'100%'} />
-				</View>
-			</View>
-		);
-	} else {
-		return (
-			<View style={styles.container}>
-				<View style={styles.headlineBox}>
-					<Text style={styles.headlineText}>ILLUMINATE YOUR FIRE!</Text>
-				</View>
-				<View style={styles.fire}>
-					<TorchOff width={'100%'} height={'100%'} />
-				</View>
-			</View>
-		);
-	}
-};
+const FireView = (props: {fire: boolean}) => (
+	<View style={styles.container}>
+		<View style={styles.headlineBox}>
+			<Text style={styles.headlineText}>
+				{props.fire ? 'SHARE YOUR FIRE!' : 'ILLUMINATE YOUR FIRE!'}
+			</Text>
+		</View>
+		<View style={styles.fire}>
+			{props.fire ? (
+				<Torch width={'100%'} height={'100%'} />
+			) : (
+				<TorchOff width={'100%'} height={'100%'} />
+			)}
+		</View>
+	</View>
+);
 
 const styles = StyleSheet.create({
 	container: {
